refactor: Extract constants
