--- apps/web/src/pages/get-started/types.ts
@@ -1,1 +1,0 @@
-export type OnboardingParams = 'usecase';
